[DCP] Consolidate the MLA DCP layout tests into the family metadata test
Review feedback on #33926: drop the trtllm_mla DCP decode hook test, and put
the decode layout properties in TestDcpDecodeLayout in
test_trtllm_mla_family_dcp_metadata.py rather than adding a standalone file.

Only the two properties not already covered are kept: the per-rank lengths
summing to the global length, and a decode step growing exactly one rank. The
third case duplicated TestGetDcpLens.test_start_none_matches_owner_count, and
those two subsume it anyway -- the partition sum pins the base case at zero
and the growth check pins rank (L-1) % n at every step, which determines the
per-rank counts by induction.

The new class sits outside the CUDA skipUnless that guards the metadata cases,
since the length math needs no GPU. This does shift the two checks off the
base-a CPU shard onto the base-b 4-gpu-b200 suite the file registers for;
get_dcp_lens keeps CPU coverage through the existing TestGetDcpLens cases.
